+++ server/config/config.exs
@@ -32,6 +32,19 @@ config :arena, ArenaWeb.Endpoint,
   pubsub_server: Arena.PubSub,
   live_view: [signing_salt: "XED/NEZq"]
 
+######################################
+# App configuration: ao_tcp_gateway #
+######################################
+
+config :ao_tcp_gateway,
+  hogar_travel_delay_ms: %{
+    gm: 5_000,
+    normal: 10_000,
+    adventurer: 10_000,
+    hero: 10_000,
+    legend: 10_000
+  }
+
 ###################################
 # App configuration: game_backend #
 ###################################